refactor: Update values-eks-example.yaml and show-env microservice configuration
This commit updates the values-eks-example.yaml file to disable the cert-manager and backstage helm charts. It also updates the show-env microservice configuration by setting the ingress host to "*.eks.kuberise.dev" and the ingress TLS secret name to "wildcard-cert-tls". Additionally, the image tag is set to "latest". These changes aim to improve the configuration and functionality of the eks-example platform.