Refactor secrets-manager values.yaml to comment out application entries
This update modifies the values.yaml file for the secrets-manager component by commenting out the application entries for 'grafana' and 'pgadmin'. This change helps to declutter the configuration and may indicate a shift in focus or usage of these applications within the platform. The commented lines maintain the structure for potential future use while enhancing clarity in the current configuration.